Add Bindings for Regex Creation
Adds the funciton definitions for creating a regex and a new wrapper
struct `Regex` for creating and managing them from Rust, including an
implementation of the `Drop` trait to automatically free the `regex_t`
pointers.

This also adds an `Error` struct with an implementaiton for the `Debug`
trait which prints out the formatted error string from Oniguruma.