Initial Work on Pattern Implementation
This allows the `onig::Regex` to be used with the `std::str::pattern`
API. `Pattern`s can be used to find a list of matches within a string,
to find prefixes, to check the existence of a substring and to trim
strings.

This implementation is probably a bit rough around the edges. I'd have
liked to have used the existing `RegexMatchs` iterator but couldn't due
to the ownership requirements of the `Pattern` trait.

Still TODO:

 * Work getting this to build on stable. Will prbbably require a feature
   switch.
 * See how well we handle empty patterns when queried for matches.