Support for `onig_scan`
Initial work on #24, for the 'scan' API. Works well enougt to implement
the `scan` example from the Oniguruma samples directory (#21).

The only niggle about this is that we have to clone the region that is
being passed back in the callback to turn it into an owned object. Maybe
we should make `onig::Region` an enum with two variants for owned and
borrowed regions?