Add MSVC Support (#36)
If we are building with the MSVC toolchain then use the existing NMAKE files instead of Cmake.

MSVC builds rely on Oniguruma's NMAKE support, not Cmake. No need to
depend on the crate for those builds.

AppVeyor triggers builds for PRs and branches. Updated the config to
skip the branch build if there's a pull request open for it.

Remove extra Travis builds by disabling builds for all but `master`.