Add Minimum Supported Rust Version Policy
This commit adds a policy of rust version support to the
README. Increasing the verion of the compiler needed to compile the
library should result in a minor version bump in the code. Clients who
require a stable version policy can therefore target `4.2` and any
patch versoin supplied by Cargo should compile with `1.27.0`.

Version 4.3.1 bumped compiler versions. This commit rolls back to
supporting 1.27.0 as it is a trivial fix by sqaushing warnings from no
clippy and fixing the error in the cast's path.

We also add Rust 1.27.0 to the Travis Build to make sure this doesn't
break silently in the future.