Replace `lazy_static` with `once_cell` (#169)
`onig` is the only dependency left in Artichoke's `Cargo.lock` that uses
`lazy_static`. It looks like most of the ecosystem has moved to
`once_cell`. `once_cell` looks to be more actively maintained and is IMO
nicer to read since there are no macros.

See also this related change in `wasm-bindgen`:

- https://github.com/rustwasm/wasm-bindgen/pull/2962

`once_cell`'s MSRV is documented to be 1.36.0:

https://github.com/matklad/once_cell/blob/eda22cec55e9d37b16d408b7d0a5b396c7feb44c/CHANGELOG.md#151