Upgrade to bitflags v2
Prepare for v6.5.0 release.

As a compatibility note, bitflags v2 does not generate a
`from_bits_unchecked` method. See the "major changes" section of the
release notes: https://github.com/bitflags/bitflags/releases/tag/2.0.0

I consider this breakage to be minor and not worth a major version bump,
but I'm happy to rev the `onig` crate to 7.0.0 if you'd prefer.

I'd appreciate getting this released. `onig` is one of the last crates
in my lockfile that hasn't yet upgraded to `bitflags` v2.